requirements_gradio.txt
torch
transformers
numpy
PyYAML
Pillow>=6.2.2
sentence-transformers==2.3.0
gradio<4.0
#gradio_imageslider==0.0.18